Instrument_class_isolation
